Move compactionJobStatusStore mocks into specific test
