+++ java/clients/src/test/java/sleeper/clients/admin/CompactionStatusReportTest.java
@@ -21,9 +21,15 @@ import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 import sleeper.clients.admin.testutils.AdminClientMockStoreBase;
+import sleeper.compaction.job.CompactionJobTestDataHelper;
+import sleeper.compaction.job.status.CompactionJobStatus;
 import sleeper.configuration.properties.InstanceProperties;
 
+import java.time.Instant;
+import java.util.List;
+
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.Mockito.when;
 import static sleeper.clients.admin.testutils.ExpectedAdminConsoleValues.COMPACTION_JOB_STATUS_REPORT_OPTION;
 import static sleeper.clients.admin.testutils.ExpectedAdminConsoleValues.COMPACTION_STATUS_REPORT_OPTION;
 import static sleeper.clients.admin.testutils.ExpectedAdminConsoleValues.EXIT_OPTION;
@@ -31,16 +37,25 @@ import static sleeper.clients.admin.testutils.ExpectedAdminConsoleValues.JOB_QUE
 import static sleeper.clients.admin.testutils.ExpectedAdminConsoleValues.JOB_QUERY_UNKNOWN_OPTION;
 import static sleeper.clients.admin.testutils.ExpectedAdminConsoleValues.MAIN_SCREEN;
 import static sleeper.clients.admin.testutils.ExpectedAdminConsoleValues.PROMPT_RETURN_TO_MAIN;
+import static sleeper.compaction.job.CompactionJobStatusTestData.finishedCompactionRun;
+import static sleeper.compaction.job.CompactionJobStatusTestData.jobCreated;
+import static sleeper.compaction.job.CompactionJobStatusTestData.startedCompactionRun;
+import static sleeper.configuration.properties.UserDefinedInstanceProperty.ID;
 import static sleeper.console.ConsoleOutput.CLEAR_CONSOLE;
+import static sleeper.core.record.process.RecordsProcessedSummaryTestData.summary;
 
 class CompactionStatusReportTest extends AdminClientMockStoreBase {
     @Nested
     @DisplayName("Compaction job status report")
     class CompactionJobStatusReport {
+        private final CompactionJobTestDataHelper dataHelper = new CompactionJobTestDataHelper();
+
         @Test
         void shouldRunCompactionJobStatusReportWithQueryTypeAll() {
             // Given
             createCompactionStatusStore();
+            when(compactionJobStatusStore.getAllJobs("test-table"))
+                    .thenReturn(exampleJobStatuses(dataHelper));
             in.enterNextPrompts(COMPACTION_STATUS_REPORT_OPTION,
                     COMPACTION_JOB_STATUS_REPORT_OPTION, "test-table", JOB_QUERY_ALL_OPTION,
                     EXIT_OPTION);
@@ -51,18 +66,18 @@ class CompactionStatusReportTest extends AdminClientMockStoreBase {
                     .endsWith(PROMPT_RETURN_TO_MAIN + CLEAR_CONSOLE + MAIN_SCREEN)
                     .contains("Compaction Job Status Report")
                     .contains("" +
-                            "Total jobs: 1\n" +
-                            "\n" +
-                            "Total standard jobs: 1\n" +
+                            "Total standard jobs: 2\n" +
                             "Total standard jobs pending: 0\n" +
                             "Total standard jobs in progress: 1\n" +
-                            "Total standard jobs finished: 0");
+                            "Total standard jobs finished: 1");
         }
 
         @Test
         void shouldRunCompactionJobStatusReportWithQueryTypeUnknown() {
             // Given
             createCompactionStatusStore();
+            when(compactionJobStatusStore.getUnfinishedJobs("test-table"))
+                    .thenReturn(exampleJobStatuses(dataHelper));
             in.enterNextPrompts(COMPACTION_STATUS_REPORT_OPTION,
                     COMPACTION_JOB_STATUS_REPORT_OPTION, "test-table", JOB_QUERY_UNKNOWN_OPTION,
                     EXIT_OPTION);
@@ -73,16 +88,28 @@ class CompactionStatusReportTest extends AdminClientMockStoreBase {
                     .endsWith(PROMPT_RETURN_TO_MAIN + CLEAR_CONSOLE + MAIN_SCREEN)
                     .contains("Compaction Job Status Report")
                     .contains("" +
-                            "Total unfinished jobs: 1\n" +
-                            "Total unfinished jobs in progress: 1\n" +
+                            "Total unfinished jobs: 2\n" +
+                            "Total unfinished jobs in progress: 2\n" +
                             "Total unfinished jobs not started: 0");
         }
     }
 
     private void createCompactionStatusStore() {
         InstanceProperties properties = createValidInstanceProperties();
-        setInstanceProperties(properties);
-        setCompactionStatusStore(properties,
-                createValidTableProperties(properties, "test-table"));
+        when(store.loadCompactionJobStatusStore(properties.get(ID)))
+                .thenReturn(compactionJobStatusStore);
+    }
+
+    private List<CompactionJobStatus> exampleJobStatuses(CompactionJobTestDataHelper dataHelper) {
+        return List.of(
+                jobCreated(dataHelper.singleFileCompaction(),
+                        Instant.parse("2023-03-15T17:52:12.001Z"),
+                        startedCompactionRun("test-task-1", Instant.parse("2023-03-15T17:53:12.001Z"))),
+                jobCreated(dataHelper.singleFileCompaction(),
+                        Instant.parse("2023-03-15T18:52:12.001Z"),
+                        finishedCompactionRun("test-task-2", summary(
+                                Instant.parse("2023-03-15T18:52:12.001Z"),
+                                Instant.parse("2023-03-15T18:53:12.001Z"),
+                                1000L, 2000L))));
     }
 }
